+++ include/game.h
@@ -6,7 +6,7 @@
 /*   By: ootsuboyoshiyuki <ootsuboyoshiyuki@stud    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/15 14:34:31 by ootsuboyosh       #+#    #+#             */
-/*   Updated: 2024/11/21 16:27:01 by ootsuboyosh      ###   ########.fr       */
+/*   Updated: 2024/11/21 18:36:06 by ootsuboyosh      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,5 +41,6 @@ void	render_scene(t_game *game);
 void	draw_wall(t_game *game, t_ray *ray, int x, int draw_start,
 			int draw_end);
 void	draw_floor_and_ceiling(t_game *game);
+int		exit_program(t_game *game);
 
 #endif
